move existing code into package and add .gitignore
